move some api_description functions into new schema_util
