+++ dropshot/src/api_description.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Oxide Computer Company
+// Copyright 2023 Oxide Computer Company
 //! Describes the endpoints and handler functions in your API
 
 use crate::handler::HttpHandlerFunc;
@@ -8,6 +8,7 @@ use crate::handler::RouteHandler;
 use crate::router::route_path_to_segments;
 use crate::router::HttpRouter;
 use crate::router::PathSegment;
+use crate::schema_util::j2oas_schema;
 use crate::server::ServerContext;
 use crate::type_util::type_is_scalar;
 use crate::type_util::type_is_string_enum;
@@ -927,440 +928,6 @@ fn is_empty(schema: &schemars::schema::Schema) -> bool {
     false
 }
 
-/// Convert from JSON Schema into OpenAPI.
-// TODO Initially this seemed like it was going to be a win, but the versions
-// of JSON Schema that the schemars and openapiv3 crates adhere to are just
-// different enough to make the conversion a real pain in the neck. A better
-// approach might be a derive(OpenAPI)-like thing, or even a generic
-// derive(schema) that we could then marshall into OpenAPI.
-// The schemars crate also seems a bit inflexible when it comes to how the
-// schema is generated wrt references vs. inline types.
-fn j2oas_schema(
-    name: Option<&String>,
-    schema: &schemars::schema::Schema,
-) -> openapiv3::ReferenceOr<openapiv3::Schema> {
-    match schema {
-        // The permissive, "match anything" schema. We'll typically see this
-        // when consumers use a type such as serde_json::Value.
-        schemars::schema::Schema::Bool(true) => {
-            openapiv3::ReferenceOr::Item(openapiv3::Schema {
-                schema_data: openapiv3::SchemaData::default(),
-                schema_kind: openapiv3::SchemaKind::Any(
-                    openapiv3::AnySchema::default(),
-                ),
-            })
-        }
-        schemars::schema::Schema::Bool(false) => {
-            panic!("We don't expect to see a schema that matches the null set")
-        }
-        schemars::schema::Schema::Object(obj) => j2oas_schema_object(name, obj),
-    }
-}
-
-fn j2oas_schema_object(
-    name: Option<&String>,
-    obj: &schemars::schema::SchemaObject,
-) -> openapiv3::ReferenceOr<openapiv3::Schema> {
-    if let Some(reference) = &obj.reference {
-        return openapiv3::ReferenceOr::Reference {
-            reference: reference.clone(),
-        };
-    }
-
-    let ty = match &obj.instance_type {
-        Some(schemars::schema::SingleOrVec::Single(ty)) => Some(ty.as_ref()),
-        Some(schemars::schema::SingleOrVec::Vec(_)) => {
-            panic!(
-                "a type array is unsupported by openapiv3:\n{}",
-                serde_json::to_string_pretty(obj)
-                    .unwrap_or_else(|_| "<can't serialize>".to_string())
-            )
-        }
-        None => None,
-    };
-
-    let kind = match (ty, &obj.subschemas) {
-        (Some(schemars::schema::InstanceType::Null), None) => {
-            openapiv3::SchemaKind::Type(openapiv3::Type::String(
-                openapiv3::StringType {
-                    enumeration: vec![None],
-                    ..Default::default()
-                },
-            ))
-        }
-        (Some(schemars::schema::InstanceType::Boolean), None) => {
-            openapiv3::SchemaKind::Type(openapiv3::Type::Boolean {})
-        }
-        (Some(schemars::schema::InstanceType::Object), None) => {
-            j2oas_object(&obj.object)
-        }
-        (Some(schemars::schema::InstanceType::Array), None) => {
-            j2oas_array(&obj.array)
-        }
-        (Some(schemars::schema::InstanceType::Number), None) => {
-            j2oas_number(&obj.format, &obj.number, &obj.enum_values)
-        }
-        (Some(schemars::schema::InstanceType::String), None) => {
-            j2oas_string(&obj.format, &obj.string, &obj.enum_values)
-        }
-        (Some(schemars::schema::InstanceType::Integer), None) => {
-            j2oas_integer(&obj.format, &obj.number, &obj.enum_values)
-        }
-        (None, Some(subschema)) => j2oas_subschemas(subschema),
-        (None, None) => {
-            openapiv3::SchemaKind::Any(openapiv3::AnySchema::default())
-        }
-        (Some(_), Some(_)) => panic!(
-            "a schema can't have both a type and subschemas:\n{}",
-            serde_json::to_string_pretty(&obj)
-                .unwrap_or_else(|_| "<can't serialize>".to_string())
-        ),
-    };
-
-    let mut data = openapiv3::SchemaData::default();
-
-    if matches!(
-        &obj.extensions.get("nullable"),
-        Some(serde_json::Value::Bool(true))
-    ) {
-        data.nullable = true;
-    }
-
-    if let Some(metadata) = &obj.metadata {
-        data.title = metadata.title.clone();
-        data.description = metadata.description.clone();
-        data.default = metadata.default.clone();
-        data.deprecated = metadata.deprecated;
-        data.read_only = metadata.read_only;
-        data.write_only = metadata.write_only;
-    }
-
-    if let Some(name) = name {
-        data.title = Some(name.clone());
-    }
-    if let Some(example) = obj.extensions.get("example") {
-        data.example = Some(example.clone());
-    }
-
-    openapiv3::ReferenceOr::Item(openapiv3::Schema {
-        schema_data: data,
-        schema_kind: kind,
-    })
-}
-
-fn j2oas_subschemas(
-    subschemas: &schemars::schema::SubschemaValidation,
-) -> openapiv3::SchemaKind {
-    match (
-        &subschemas.all_of,
-        &subschemas.any_of,
-        &subschemas.one_of,
-        &subschemas.not,
-    ) {
-        (Some(all_of), None, None, None) => openapiv3::SchemaKind::AllOf {
-            all_of: all_of
-                .iter()
-                .map(|schema| j2oas_schema(None, schema))
-                .collect::<Vec<_>>(),
-        },
-        (None, Some(any_of), None, None) => openapiv3::SchemaKind::AnyOf {
-            any_of: any_of
-                .iter()
-                .map(|schema| j2oas_schema(None, schema))
-                .collect::<Vec<_>>(),
-        },
-        (None, None, Some(one_of), None) => openapiv3::SchemaKind::OneOf {
-            one_of: one_of
-                .iter()
-                .map(|schema| j2oas_schema(None, schema))
-                .collect::<Vec<_>>(),
-        },
-        (None, None, None, Some(not)) => openapiv3::SchemaKind::Not {
-            not: Box::new(j2oas_schema(None, not)),
-        },
-        _ => panic!("invalid subschema {:#?}", subschemas),
-    }
-}
-
-fn j2oas_integer(
-    format: &Option<String>,
-    number: &Option<Box<schemars::schema::NumberValidation>>,
-    enum_values: &Option<Vec<serde_json::value::Value>>,
-) -> openapiv3::SchemaKind {
-    let format = match format.as_ref().map(|s| s.as_str()) {
-        None => openapiv3::VariantOrUnknownOrEmpty::Empty,
-        Some("int32") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::IntegerFormat::Int32,
-        ),
-        Some("int64") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::IntegerFormat::Int64,
-        ),
-        Some(other) => {
-            openapiv3::VariantOrUnknownOrEmpty::Unknown(other.to_string())
-        }
-    };
-
-    let (multiple_of, minimum, exclusive_minimum, maximum, exclusive_maximum) =
-        match number {
-            None => (None, None, false, None, false),
-            Some(number) => {
-                let multiple_of = number.multiple_of.map(|f| f as i64);
-                let (minimum, exclusive_minimum) =
-                    match (number.minimum, number.exclusive_minimum) {
-                        (None, None) => (None, false),
-                        (Some(f), None) => (Some(f as i64), false),
-                        (None, Some(f)) => (Some(f as i64), true),
-                        _ => panic!("invalid"),
-                    };
-                let (maximum, exclusive_maximum) =
-                    match (number.maximum, number.exclusive_maximum) {
-                        (None, None) => (None, false),
-                        (Some(f), None) => (Some(f as i64), false),
-                        (None, Some(f)) => (Some(f as i64), true),
-                        _ => panic!("invalid"),
-                    };
-
-                (
-                    multiple_of,
-                    minimum,
-                    exclusive_minimum,
-                    maximum,
-                    exclusive_maximum,
-                )
-            }
-        };
-
-    let enumeration = enum_values
-        .iter()
-        .flat_map(|v| {
-            v.iter().map(|vv| match vv {
-                serde_json::Value::Null => None,
-                serde_json::Value::Number(value) => {
-                    Some(value.as_i64().unwrap())
-                }
-                _ => panic!("unexpected enumeration value {:?}", vv),
-            })
-        })
-        .collect::<Vec<_>>();
-
-    openapiv3::SchemaKind::Type(openapiv3::Type::Integer(
-        openapiv3::IntegerType {
-            format,
-            multiple_of,
-            exclusive_minimum,
-            exclusive_maximum,
-            minimum,
-            maximum,
-            enumeration,
-        },
-    ))
-}
-
-fn j2oas_number(
-    format: &Option<String>,
-    number: &Option<Box<schemars::schema::NumberValidation>>,
-    enum_values: &Option<Vec<serde_json::value::Value>>,
-) -> openapiv3::SchemaKind {
-    let format = match format.as_ref().map(|s| s.as_str()) {
-        None => openapiv3::VariantOrUnknownOrEmpty::Empty,
-        Some("float") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::NumberFormat::Float,
-        ),
-        Some("double") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::NumberFormat::Double,
-        ),
-        Some(other) => {
-            openapiv3::VariantOrUnknownOrEmpty::Unknown(other.to_string())
-        }
-    };
-
-    let (multiple_of, minimum, exclusive_minimum, maximum, exclusive_maximum) =
-        match number {
-            None => (None, None, false, None, false),
-            Some(number) => {
-                let multiple_of = number.multiple_of;
-                let (minimum, exclusive_minimum) =
-                    match (number.minimum, number.exclusive_minimum) {
-                        (None, None) => (None, false),
-                        (s @ Some(_), None) => (s, false),
-                        (None, s @ Some(_)) => (s, true),
-                        _ => panic!("invalid"),
-                    };
-                let (maximum, exclusive_maximum) =
-                    match (number.maximum, number.exclusive_maximum) {
-                        (None, None) => (None, false),
-                        (s @ Some(_), None) => (s, false),
-                        (None, s @ Some(_)) => (s, true),
-                        _ => panic!("invalid"),
-                    };
-
-                (
-                    multiple_of,
-                    minimum,
-                    exclusive_minimum,
-                    maximum,
-                    exclusive_maximum,
-                )
-            }
-        };
-
-    let enumeration = enum_values
-        .iter()
-        .flat_map(|v| {
-            v.iter().map(|vv| match vv {
-                serde_json::Value::Null => None,
-                serde_json::Value::Number(value) => {
-                    Some(value.as_f64().unwrap())
-                }
-                _ => panic!("unexpected enumeration value {:?}", vv),
-            })
-        })
-        .collect::<Vec<_>>();
-
-    openapiv3::SchemaKind::Type(openapiv3::Type::Number(
-        openapiv3::NumberType {
-            format,
-            multiple_of,
-            exclusive_minimum,
-            exclusive_maximum,
-            minimum,
-            maximum,
-            enumeration,
-        },
-    ))
-}
-
-fn j2oas_string(
-    format: &Option<String>,
-    string: &Option<Box<schemars::schema::StringValidation>>,
-    enum_values: &Option<Vec<serde_json::value::Value>>,
-) -> openapiv3::SchemaKind {
-    let format = match format.as_ref().map(|s| s.as_str()) {
-        None => openapiv3::VariantOrUnknownOrEmpty::Empty,
-        Some("date") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::StringFormat::Date,
-        ),
-        Some("date-time") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::StringFormat::DateTime,
-        ),
-        Some("password") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::StringFormat::Password,
-        ),
-        Some("byte") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::StringFormat::Byte,
-        ),
-        Some("binary") => openapiv3::VariantOrUnknownOrEmpty::Item(
-            openapiv3::StringFormat::Binary,
-        ),
-        Some(other) => {
-            openapiv3::VariantOrUnknownOrEmpty::Unknown(other.to_string())
-        }
-    };
-
-    let (max_length, min_length, pattern) = match string.as_ref() {
-        None => (None, None, None),
-        Some(string) => (
-            string.max_length.map(|n| n as usize),
-            string.min_length.map(|n| n as usize),
-            string.pattern.clone(),
-        ),
-    };
-
-    let enumeration = enum_values
-        .iter()
-        .flat_map(|v| {
-            v.iter().map(|vv| match vv {
-                serde_json::Value::Null => None,
-                serde_json::Value::String(s) => Some(s.clone()),
-                _ => panic!("unexpected enumeration value {:?}", vv),
-            })
-        })
-        .collect::<Vec<_>>();
-
-    openapiv3::SchemaKind::Type(openapiv3::Type::String(
-        openapiv3::StringType {
-            format,
-            pattern,
-            enumeration,
-            min_length,
-            max_length,
-        },
-    ))
-}
-
-fn j2oas_array(
-    array: &Option<Box<schemars::schema::ArrayValidation>>,
-) -> openapiv3::SchemaKind {
-    let arr = array.as_ref().unwrap();
-
-    openapiv3::SchemaKind::Type(openapiv3::Type::Array(openapiv3::ArrayType {
-        items: match &arr.items {
-            Some(schemars::schema::SingleOrVec::Single(schema)) => {
-                Some(box_reference_or(j2oas_schema(None, &schema)))
-            }
-            Some(schemars::schema::SingleOrVec::Vec(_)) => {
-                panic!("OpenAPI v3.0.x cannot support tuple-like arrays")
-            }
-            None => None,
-        },
-        min_items: arr.min_items.map(|n| n as usize),
-        max_items: arr.max_items.map(|n| n as usize),
-        unique_items: arr.unique_items.unwrap_or(false),
-    }))
-}
-
-fn box_reference_or<T>(
-    r: openapiv3::ReferenceOr<T>,
-) -> openapiv3::ReferenceOr<Box<T>> {
-    match r {
-        openapiv3::ReferenceOr::Item(schema) => {
-            openapiv3::ReferenceOr::boxed_item(schema)
-        }
-        openapiv3::ReferenceOr::Reference { reference } => {
-            openapiv3::ReferenceOr::Reference { reference }
-        }
-    }
-}
-
-fn j2oas_object(
-    object: &Option<Box<schemars::schema::ObjectValidation>>,
-) -> openapiv3::SchemaKind {
-    match object {
-        None => openapiv3::SchemaKind::Type(openapiv3::Type::Object(
-            openapiv3::ObjectType::default(),
-        )),
-        Some(obj) => openapiv3::SchemaKind::Type(openapiv3::Type::Object(
-            openapiv3::ObjectType {
-                properties: obj
-                    .properties
-                    .iter()
-                    .map(|(prop, schema)| {
-                        (
-                            prop.clone(),
-                            box_reference_or(j2oas_schema(None, schema)),
-                        )
-                    })
-                    .collect::<_>(),
-                required: obj.required.iter().cloned().collect::<_>(),
-                additional_properties: obj.additional_properties.as_ref().map(
-                    |schema| match schema.as_ref() {
-                        schemars::schema::Schema::Bool(b) => {
-                            openapiv3::AdditionalProperties::Any(*b)
-                        }
-                        schemars::schema::Schema::Object(obj) => {
-                            openapiv3::AdditionalProperties::Schema(Box::new(
-                                j2oas_schema_object(None, obj),
-                            ))
-                        }
-                    },
-                ),
-                min_properties: obj.min_properties.map(|n| n as usize),
-                max_properties: obj.max_properties.map(|n| n as usize),
-            },
-        )),
-    }
-}
-
 /// This object is used to specify configuration for building an OpenAPI
 /// definition document.  It is constructed using [`ApiDescription::openapi()`].
 /// Additional optional properties may be added and then the OpenAPI definition
@@ -1554,8 +1121,6 @@ impl Default for ExtensionMode {
 
 #[cfg(test)]
 mod test {
-    use super::j2oas_schema;
-    use crate::api_description::j2oas_schema_object;
     use crate::endpoint;
     use crate::error::HttpError;
     use crate::handler::RequestContext;
@@ -1665,59 +1230,6 @@ mod test {
     }
 
     #[test]
-    fn test_empty_struct() {
-        #[derive(JsonSchema)]
-        struct Empty {}
-
-        let settings = schemars::gen::SchemaSettings::openapi3();
-        let mut generator = schemars::gen::SchemaGenerator::new(settings);
-
-        let schema = Empty::json_schema(&mut generator);
-        let _ = j2oas_schema(None, &schema);
-    }
-
-    #[test]
-    fn test_garbage_barge_structure_conversion() {
-        #[allow(dead_code)]
-        #[derive(JsonSchema)]
-        struct SuperGarbage {
-            string: String,
-            strings: Vec<String>,
-            more_strings: [String; 3],
-            substruct: Substruct,
-            more: Option<Substruct>,
-            union: Union,
-            map: std::collections::BTreeMap<String, String>,
-        }
-
-        #[allow(dead_code)]
-        #[derive(JsonSchema)]
-        struct Substruct {
-            ii32: i32,
-            uu64: u64,
-            ff: f32,
-            dd: f64,
-            b: bool,
-        }
-
-        #[allow(dead_code)]
-        #[derive(JsonSchema)]
-        enum Union {
-            A { a: u32 },
-            B { b: f32 },
-        }
-
-        let settings = schemars::gen::SchemaSettings::openapi3();
-        let mut generator = schemars::gen::SchemaGenerator::new(settings);
-
-        let schema = SuperGarbage::json_schema(&mut generator);
-        let _ = j2oas_schema(None, &schema);
-        for (key, schema) in generator.definitions().iter() {
-            let _ = j2oas_schema(Some(key), schema);
-        }
-    }
-
-    #[test]
     fn test_two_bodies() {
         #[derive(Deserialize, JsonSchema)]
         struct AStruct {}
@@ -1772,52 +1284,6 @@ mod test {
             error,
             "the parameter 'thing' is specified for both query and path \
              parameters",
-        );
-    }
-
-    #[test]
-    fn test_additional_properties() {
-        #[allow(dead_code)]
-        #[derive(JsonSchema)]
-        enum Union {
-            A { a: u32 },
-        }
-        let settings = schemars::gen::SchemaSettings::openapi3();
-        let mut generator = schemars::gen::SchemaGenerator::new(settings);
-        let schema = Union::json_schema(&mut generator);
-        let _ = j2oas_schema(None, &schema);
-        for (key, schema) in generator.definitions().iter() {
-            let _ = j2oas_schema(Some(key), schema);
-        }
-    }
-
-    #[test]
-    fn test_nullable() {
-        #[allow(dead_code)]
-        #[derive(JsonSchema)]
-        struct Foo {
-            bar: String,
-        }
-        let settings = schemars::gen::SchemaSettings::openapi3();
-        let generator = schemars::gen::SchemaGenerator::new(settings);
-        let root_schema = generator.into_root_schema_for::<Option<Foo>>();
-        let schema = root_schema.schema;
-        let os = j2oas_schema_object(None, &schema);
-
-        assert_eq!(
-            os,
-            openapiv3::ReferenceOr::Item(openapiv3::Schema {
-                schema_data: openapiv3::SchemaData {
-                    title: Some("Nullable_Foo".to_string()),
-                    nullable: true,
-                    ..Default::default()
-                },
-                schema_kind: openapiv3::SchemaKind::AllOf {
-                    all_of: vec![openapiv3::ReferenceOr::Reference {
-                        reference: "#/components/schemas/Foo".to_string()
-                    }],
-                },
-            })
         );
     }
 
@@ -1938,49 +1404,5 @@ mod test {
                 .into_iter()
                 .collect::<HashSet<_>>()
         )
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_bad_schema() {
-        #![allow(unused)]
-
-        #[derive(JsonSchema)]
-        #[schemars(tag = "which")]
-        enum Which {
-            This,
-            That,
-        }
-
-        #[derive(JsonSchema)]
-        struct BlackSheep {
-            #[schemars(flatten)]
-            you_can_get_with: Which,
-        }
-
-        let schema = schemars::schema_for!(BlackSheep).schema;
-
-        let _ = j2oas_schema_object(None, &schema);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_two_types() {
-        #![allow(unused)]
-
-        #[derive(JsonSchema)]
-        enum One {
-            One,
-        }
-
-        #[derive(JsonSchema)]
-        struct Uno {
-            #[schemars(flatten)]
-            one: One,
-        }
-
-        let schema = schemars::schema_for!(Uno).schema;
-
-        let _ = j2oas_schema_object(None, &schema);
     }
 }
